Update Game.cpp for Metapod
Added the chance of a Metapod to spawn as well as changed it so the only option the Metapod has is to harden. Could not add to the Enemy.cpp as I couldn't distinguish the Metapod from other enemy types. This is the least obstructive way i found to be able to have the Metapod have a special set of actions (or lack there of).